n+1 card game

+++ back/수학/정수론/2247.cpp
@@ -6,9 +6,9 @@ const int mod = 1000000;
 long long CSOD(int n)
 {
     long long sum=0;
-    for(int i = 2; i*i<=n; i++)
+    for(long long i = 2; i*i<=n; i++)
     {
-        int j = n/i;
+        long long j = n/i;
         sum += i * (j - i + 1) + (j - i) * (j + i + 1) / 2;
     }
     return sum%mod;
